database/connection_postgresql.py
# %% Imports
import os

from dotenv import load_dotenv
import psycopg2
from pgvector.psycopg2 import register_vector

load_dotenv()

# %%  Postgresql
# Connect to the database
connPostgres = psycopg2.connect(
    dbname=os.getenv("POSTGRES_DB"),
    user=os.getenv("POSTGRES_USER"),
    password=os.getenv("POSTGRES_PASSWORD"),
    host=os.getenv("POSTGRES_HOST"),
)
# Register the vector type with psycopg2
register_vector(connPostgres)
