requirements.txt
flask
Flask-SQLAlchemy
matplotlib
pandas
numpy
seaborn
requests
gspread
scikit-learn
openpyxl
openai
python-dotenv
tiktoken
psycopg2-binary
pgvector
gunicorn
mysql-connector-python
Flask-HTTPAuth
